refactor: Change package ids for all modules.
We should put package ids for all modules
under com.decathlon.vitamin instead of
com.decathlon.android.vitamin.
