+++ checkboxes/src/main/java/com/decathlon/vitamin/checkboxes/VitaminCheckboxes.kt
@@ -1,8 +1,7 @@
-package com.decathlon.android.checkbox
+package com.decathlon.vitamin.checkboxes
 
 import android.content.Context
 import android.util.AttributeSet
-import com.decathlon.android.checkboxes.R
 import com.google.android.material.checkbox.MaterialCheckBox
 
 open class VitaminCheckbox @JvmOverloads constructor(
