+++ foundation/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.decathlon.android.foundation" />
+<manifest package="com.decathlon.vitamin.foundation" />
