+++ sample/src/main/res/layout/fragment_switches.xml
@@ -10,7 +10,7 @@
         android:layout_height="wrap_content"
         android:padding="16dp">
 
-        <com.decathlon.android.switches.VitaminSwitch
+        <com.decathlon.vitamin.switches.VitaminSwitch
             android:id="@+id/checkedSwitch"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -20,7 +20,7 @@
             app:layout_constraintTop_toTopOf="parent"
             tools:ignore="HardcodedText" />
 
-        <com.decathlon.android.switches.VitaminSwitch
+        <com.decathlon.vitamin.switches.VitaminSwitch
             android:id="@+id/uncheckedSwitch"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -31,7 +31,7 @@
             app:layout_constraintTop_toBottomOf="@id/checkedSwitch"
             tools:ignore="HardcodedText" />
 
-        <com.decathlon.android.switches.VitaminSwitch
+        <com.decathlon.vitamin.switches.VitaminSwitch
             android:id="@+id/disabledCheckedSwitch"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -43,7 +43,7 @@
             app:layout_constraintTop_toBottomOf="@id/uncheckedSwitch"
             tools:ignore="HardcodedText" />
 
-        <com.decathlon.android.switches.VitaminSwitch
+        <com.decathlon.vitamin.switches.VitaminSwitch
             android:id="@+id/disabledUnCheckedSwitch"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
