+++ text-inputs/src/main/java/com/decathlon/vitamin/textinputs/VitaminTextInputs.kt
@@ -1,9 +1,8 @@
-package com.decathlon.android.textinputs
+package com.decathlon.vitamin.textinputs
 
 import android.content.Context
 import android.util.AttributeSet
 import com.google.android.material.textfield.TextInputLayout
-
 
 open class VitaminTextInputLayoutFilled @JvmOverloads constructor(
     context: Context,
